# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Traces App REST API
  description: |
    REST endpoints for the Traces (Taxteck) application. Blitz RPC endpoints are documented in docs/API.md.

    ## Authentication
    All endpoints except `/api/env/license` require either:
    - `Authorization: Bearer tt_{hex}` API token, or
    - Blitz session cookie (`not-track_*`) from login

    Create API tokens via `POST /api/rpc/createApiToken` after logging in.
  version: 1.0.2
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://{host}
    description: Production
    variables:
      host:
        default: example.com
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "tt_{hex}"
      description: API token created via createApiToken RPC
...
security:
  - bearerAuth: []
paths:
...
  /api/file/upload:
    post:
      tags: [Files]
      summary: Upload files to server directory
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
          description: Destination directory path on server
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Upload successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        originalName:
                          type: string
                        uploadedName:
                          type: string
                        path:
                          type: string
        '401':
          description: Unauthorized
...
  /api/file/{slug}:
    get:
      tags: [Files]
      summary: Download a file from public storage
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: array
            items:
              type: string
          style: simple
          explode: false
      responses:
        '200':
          description: File bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized